seo settings issue fixed

+++ src/Pages/SeoManagement.php
@@ -208,7 +208,7 @@ class SeoManagement extends Page implements HasTable
      {
          $avg = SeoMeta::query()->avg('seo_score');
  
-         return (int) round($avg ?? 0);
+         return (int) round((float) ($avg ?? 0));
      }
  
      /**
